fix: use adaptive download sidebar on iPad
